requirements.txt
# Core dependencies
fastapi
uvicorn
python-dotenv
python-multipart
requests
pytest

# LlamaIndex and related packages
llama-index
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-openai
llama-index-embeddings-ollama
llama-index-agent-openai
llama-index-cli

# Vector store
chromadb

# PDF processing
PyMuPDF==1.25.3

# ML/AI
openai

# Essential dependencies
python-multipart==0.0.20
